+++ SRD05VDCSL-lib/src/devices/SRD05VDCSL.h
@@ -1,0 +1,35 @@
+// Copyright (c) 2022 Roman Gorielov. All Rights Reserved.
+// 
+// This software is the confidential and proprietary information of Roman Gorielov.
+// It is furnished under license and may only be used or copied in accordance
+// with the terms of such license.
+// This software is subject to change without notice and no information
+// contained in it should be construed as commitment by Roman Gorielov.
+
+#ifndef _SRD05VDCSL_H_
+#define _SRD05VDCSL_H_
+
+#include "devices/DigitalDevices.h"
+
+struct SRD05VDCSLDevice : public RelayDevice
+{
+    SRD05VDCSLDevice() = delete;
+    SRD05VDCSLDevice(const uint8_t t_pin, IStream<uint8_t>* const t_stream)
+        : RelayDevice(RELAY_DEVICE_TYPE, t_pin, t_stream) {};
+    SRD05VDCSLDevice(const SRD05VDCSLDevice& source) noexcept
+        : RelayDevice(source) {};
+    SRD05VDCSLDevice(SRD05VDCSLDevice&& source) noexcept
+        : RelayDevice(dynamic_cast<RelayDevice&&>(source)) {};
+    virtual ~SRD05VDCSLDevice() = default;
+
+    SRD05VDCSLDevice& operator= (const SRD05VDCSLDevice& source) noexcept
+    {
+        if(this != &source)
+        {
+            RelayDevice::operator=(source);
+        }
+        return *this;
+    }        
+};
+
+#endif