+++ ZMPT101B-lib/sample/test/FakeStream.h
@@ -10,22 +10,21 @@
 #define _FAKE_STREAM_H_
 
 #include <inttypes.h>
+#include <stdlib.h>
 #include "stream/IStream.h"
 #include "FakeTimer.h"
-
-#define abs(x) ((x)>0?(x):-(x))
 
 class FakeStream : public IStream<uint16_t>
 {
     private:
-        uint16_t m_minValue;
-        uint16_t m_maxValue;
+        const uint16_t m_minValue;
+        const uint16_t m_maxValue;
         uint16_t m_stepValue;
         uint16_t m_currentValue;
         bool m_isSetToRead;
         bool m_hasError;
     public:
-        FakeStream(uint16_t t_minValue, uint16_t t_maxValue)
+        FakeStream(const uint16_t t_minValue, const uint16_t t_maxValue)
             : m_minValue(t_minValue), m_maxValue(t_maxValue), 
               m_isSetToRead(false), m_hasError(false)
         {
@@ -34,7 +33,7 @@ class FakeStream : public IStream<uint16_t>
 
         ~FakeStream() = default;
 
-        void begin(StreamMode t_mode) override
+        void begin(const StreamMode t_mode) override
         {
             m_hasError = false;
             if(t_mode == StreamMode::Read)
@@ -75,13 +74,18 @@ class FakeStream : public IStream<uint16_t>
             }
         };
 
-        void write(uint16_t t_data) override {};
+        void write(const uint16_t t_data) override {};
 
         void end() override
         {
             m_isSetToRead = false;
             m_hasError = false;
         };
+
+        uint8_t getState() override
+        {
+            return 0;
+        }
 
         bool canRead() override
         {
@@ -102,10 +106,20 @@ class FakeStream : public IStream<uint16_t>
         {
             if(hasError())
             {
-                return IO_ERROR_STREAM_CLOSED;
+                return STREAM_CLOSED_IO_ERROR;
             };
             return NO_ERROR;
         };
+
+        IStream<uint16_t>* clone() const override
+        {
+            auto stream = new FakeStream(m_minValue, m_maxValue);
+            stream->m_isSetToRead = m_isSetToRead;
+            stream->m_currentValue = m_currentValue;
+            stream->m_hasError = m_hasError;
+
+            return stream;
+        }        
 };
 
 #endif