+++ io-lib/src/stream/VoltageStream.h
@@ -23,22 +23,25 @@ class IVoltageStream : public IStream<uint16_t>
         virtual ~IVoltageStream() = default;
 
         virtual float getVoltage() = 0;
-        virtual void setPwm(int t_percentage) = 0;
+        virtual void setPwm(const int t_percentage) = 0;
 };
 
 class VoltageStream : protected AnalogStream, public IVoltageStream
 {
     public:
-        VoltageStream(IPortAdapter<int>* t_adapter) 
+        VoltageStream() = delete;
+        VoltageStream(IPortAdapter<int>* const t_adapter) 
             : AnalogStream(t_adapter) {};
 
-        ~VoltageStream() = default;
+        virtual ~VoltageStream() = default;
 
-        void begin(StreamMode t_mode) override;
+        void begin(const StreamMode t_mode) override;
         uint16_t read() override;
-        void write(uint16_t t_data) override;
+        void write(const uint16_t t_data) override;
         void end() override;
 
+        uint8_t getState() override;
+        
         bool canRead() override;
         bool canWrite() override;
         bool hasError() override;
@@ -46,7 +49,7 @@ class VoltageStream : protected AnalogStream, public IVoltageStream
         err_t getLastError() override;
 
         float getVoltage() override;
-        void setPwm(int t_percentage) override;        
+        void setPwm(const int t_percentage) override; 
 };
 
 #endif