+++ device-abstraction-lib/CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.2201]
+### Added
+- Added abstraction for Arduino Devices.
+- Added abstraction for Arduino Device's controllers.
+- Added general definitions for Arduino Devices.
+
+### Removed
+- Moved _AnalogSensor_ class to __HC_LIB.Devices__ library.
+- Removed dependency on __HC-LIB.IO__ library.
+
 ## [1.0.2111]
 ### Added
-- Added definitions of abrstactions for analog Sensors.
-- Added definitions of abstractions for Sensor's Readers.
+- Added abstractions for analog Sensors.
+- Added abstractions for Sensor's Readers.