+++ system-lib/CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.2203]
+### Added
+- Added definition of  _ICloneable_ interface. 
+
+## [1.0.2201]
+### Added
+- Added _Expected<T>_ class for error handling.
+- Added a new error code for "out of range" error.
+
+### Changed
+- The error definitions were renamed.
+
 ## [1.0.2111]
 ### Added
 - Added error definitions.